pyproject.toml
[project]
name = "nonebot-plugin-tsugu-frontend"
version = "0.1.0"
description = "基于 tsugu-python-frontend 的 tsugu-bangdream-bot NoneBot 插件"
authors = [
    {name = "zhaomaoniu", email = "2667292003@qq.com"},
]
dependencies = [
    "nonebot2>=2.2.1",
    "nonebot-plugin-alconna>=0.42.4",
    "tsugu>=0.6.0",
    "nonebot-plugin-userinfo>=0.2.4",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
